move stopwords removal logic into bleep
