+++ server/bleep/Cargo.toml
@@ -111,7 +111,6 @@ git-version = "0.3.5"
 gix = { git = "https://github.com/BloopAI/gitoxide", version="0.55.2", features = ["blocking-http-transport-reqwest-rust-tls-no-trust-dns", "pack-cache-lru-static"] }
 
 # semantic
-rake = { git = "https://github.com/rmuller-ml/rake-rs.git", branch="determism_special_characters" }
 qdrant-client = { version = "1.5.0", default-features = false }
 tiktoken-rs = "0.4.5"
 tokenizers = { version = "0.14.0", default-features = false, features = ["progressbar", "cli", "onig", "esaxx_fast"] }
