+++ server/bleep/src/query.rs
@@ -4,3 +4,4 @@ pub mod languages;
 pub mod parser;
 pub mod planner;
 pub mod ranking;
+pub mod stopwords;
